+++ innopoints/schemas/__init__.py
@@ -1,1 +1,3 @@
+from .project import *
+
 from .schemas import *  # temporary until moved all schemas
